--- a/outlines/fsm/parsing.py
+++ b/outlines/fsm/parsing.py
@@ -38,7 +38,7 @@ from lark.parsers.lalr_parser import LALR_Parser, ParseConf, ParserState, _Parse
 from outlines.fsm.regex import (
     fsm_union,
     get_sub_fsms_from_seq,
-    get_token_transitions,
+    get_token_transition_keys,
     make_deterministic_fsm,
     walk_fsm,
 )
@@ -570,7 +570,7 @@ class PartialScanner(Scanner):
 
         text_part = text[start_pos:]
 
-        text_transitions = get_token_transitions(
+        text_transitions = get_token_transition_keys(
             self.fsm.fsm_info.alphabet_symbol_mapping,
             self.fsm.fsm_info.alphabet_anything_value,
             text_part,