@M0gician 2024-05-31
Should we add some test cases like the one in #833 to verify if this `\x00` issue is solved in later updates as well? I am a little concerned whether numba's bug fix will break the current implementation in the future.