@lapp0 2024-05-31
> Should we add some test cases like the one in #833 to verify if this `\x00` issue is solved in later updates as well? I am a little concerned whether numba's bug fix will break the current implementation in the future.

You mean a test case to ensure that a `unicode_type` starting with `\x00` isn't broken in numba in the same way it breaks `UnicodeCharSeq`?