@M0gician 2024-06-01
Yes exactly

________________________________
From: Andrew Lapp ***@***.***>
Sent: Saturday, June 1, 2024 3:14:48 AM
To: outlines-dev/outlines ***@***.***>
Cc: Tommy Yang ***@***.***>; Mention ***@***.***>
Subject: Re: [outlines-dev/outlines] Fix null byte `\x00` issue in byte level fsm resulting in `KeyError` in `BetterFSM::FSMInfo`  (PR #930)


Should we add some test cases like the one in #833<https://github.com/outlines-dev/outlines/issues/833> to verify if this \x00 issue is solved in later updates as well? I am a little concerned whether numba's bug fix will break the current implementation in the future.

You mean a test case to ensure that a unicode_type starting with \x00 isn't broken in numba in the same way it breaks UnicodeCharSeq?

—
Reply to this email directly, view it on GitHub<https://github.com/outlines-dev/outlines/pull/930#issuecomment-2142847132>, or unsubscribe<https://github.com/notifications/unsubscribe-auth/ADJEHYLTECZINZ37C5OIELLZFDD2RAVCNFSM6AAAAABIRYPGICVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDCNBSHA2DOMJTGI>.
You are receiving this because you were mentioned.Message ID: ***@***.***>