@lapp0 2024-06-01
I have introduced two new tests to verify the behavior:

# `def test_numba_leading_null_byte_UnicodeCharSeq_remains_broken()`

This test asserts that https://github.com/numba/numba/issues/9542 is still a problem. If this test **fails** we can **safely use `UnicodeCharSeq` to represent byte fsms.

`UnicodeCharSeq` allows for a cleaner representation of byte fsms than `unicode_type`, so its preferable that we use it, but this test failing doesn't necessarily indicate anything is broken.

# `def test_numba_leading_null_byte_unicode_type_sane()`

This test asserts that the `unicode_type` representation of null bytes (introduced in this PR) is legal.

If this fails, then we should consider reverting to use of `UnicodeCharSeq` (assuming the previous test also fails, indicating `UnicodeCharSeq` is viable).